Move ui_backend to pkg/uibackend
This is the fifth commit in the chain for reorganising under a common
cmd and pkg directory structure. This commit moves ui_backend
package into pkg/uibackend. The underscore is dropped because golang
package names shouldn't have underscores in them.
