+++ pkg/uibackend/api/client/client.cfg.yaml
@@ -1,7 +1,7 @@
 package: client
 output: client/client.gen.go
 additional-imports:
-  - package: github.com/openclarity/vmclarity/ui_backend/api/models
+  - package: github.com/openclarity/vmclarity/pkg/uibackend/api/models
     alias: .
 generate:
   client: true
