+++ pkg/uibackend/rest/dashboard_findings_impact.go
@@ -28,7 +28,7 @@ import (
 	backendmodels "github.com/openclarity/vmclarity/api/models"
 	"github.com/openclarity/vmclarity/pkg/shared/findingkey"
 	"github.com/openclarity/vmclarity/pkg/shared/utils"
-	"github.com/openclarity/vmclarity/ui_backend/api/models"
+	"github.com/openclarity/vmclarity/pkg/uibackend/api/models"
 )
 
 const (
