+++ pkg/uibackend/rest/dashboard_findings_impact_test.go
@@ -26,7 +26,7 @@ import (
 	backendmodels "github.com/openclarity/vmclarity/api/models"
 	"github.com/openclarity/vmclarity/pkg/shared/findingkey"
 	"github.com/openclarity/vmclarity/pkg/shared/utils"
-	"github.com/openclarity/vmclarity/ui_backend/api/models"
+	"github.com/openclarity/vmclarity/pkg/uibackend/api/models"
 )
 
 func Test_processFindings(t *testing.T) {
