+++ pkg/uibackend/rest/dashboard_findings_trends_test.go
@@ -22,7 +22,7 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"gotest.tools/v3/assert"
 
-	"github.com/openclarity/vmclarity/ui_backend/api/models"
+	"github.com/openclarity/vmclarity/pkg/uibackend/api/models"
 )
 
 func Test_createTimes(t *testing.T) {
