+++ pkg/uibackend/rest/dashboard_riskiest_assets_test.go
@@ -25,7 +25,7 @@ import (
 
 	backendmodels "github.com/openclarity/vmclarity/api/models"
 	"github.com/openclarity/vmclarity/pkg/shared/utils"
-	"github.com/openclarity/vmclarity/ui_backend/api/models"
+	"github.com/openclarity/vmclarity/pkg/uibackend/api/models"
 )
 
 func Test_getTotalFindingField(t *testing.T) {
